Registration & Newsletter notifications improvements
