+++ layouts/basic/modules/Settings/SystemWarnings/YetiForce/Registration.tpl
@@ -1,0 +1,20 @@
+{*<!-- {[The file is published on the basis of YetiForce Public License 3.0 that can be found in the following directory: licenses/LicenseEN.txt or yetiforce.com]} -->*}
+{strip}
+	<h6 class="h3">
+		{App\Language::translate('LBL_REGISTRATION','Settings:SystemWarnings')}
+	</h6>
+	<p>{App\Language::translate('LBL_YETIFORCE_REGISTRATION_ERROR')}</p>
+	<form class="validateForm" method="post" action="index.php">
+		<div class="float-right mr-2">
+			<a href="index.php?module=Companies&parent=Settings&view=List&displayModal=online" target="_blank"
+			   class="btn btn-success mr-1">
+				<span class="fas fa-check mr-1"></span>
+				{App\Language::translate('LBL_COMPANY_DATA','Settings:SystemWarnings')}
+			</a>
+			<button type="button" class="btn btn-danger cancel">
+				<span class="fas fa-ban mr-1"></span>
+				{App\Language::translate('LBL_REMIND_LATER','Settings:SystemWarnings')}
+			</button>
+		</div>
+	</form>
+{/strip}
