+++ layouts/basic/modules/Vtiger/Footer.tpl
@@ -65,10 +65,10 @@
 						{if !\App\YetiForce\Register::verify(true)}
 							<li class="page-item u-cursor-pointer">
 								<a class="page-link text-danger js-popover-tooltip" role="button"
-								   data-content="{\App\Language::translate('LBL_YETIFORCE_REGISTRATION_ERROR', $MODULE_NAME)}"
+								   data-content="{\App\Purifier::encodeHtml(\App\Language::translate('LBL_YETIFORCE_REGISTRATION_ERROR', $MODULE_NAME))}"
 								   title="{\App\Language::translate('LBL_YETIFORCE_REGISTRATION', $MODULE_NAME)}"
 										{if $USER_MODEL->isAdminUser()}
-											href="index.php?parent=Settings&module=Companies&view=List"
+											href="index.php?parent=Settings&module=Companies&view=List&displayModal=online"
 										{else}
 											href="#"
 										{/if} >
